Move the InventoryRepository to the Repository dir
